+++ crates/biome_js_formatter/src/js/expressions/call_arguments.rs
@@ -1,4 +1,5 @@
 use crate::context::trailing_comma::FormatTrailingComma;
+use crate::js::bindings::parameters::has_only_simple_parameters;
 use crate::js::declarations::function_declaration::FormatFunctionOptions;
 use crate::js::expressions::arrow_function_expression::{
     is_multiline_template_starting_on_same_line, FormatJsArrowFunctionExpressionOptions,
@@ -359,7 +360,7 @@ fn write_grouped_arguments(
                 grouped_arg.cache_function_body(f);
             }
             AnyJsCallArgument::AnyJsExpression(AnyJsExpression::JsFunctionExpression(function))
-                if !other_args.is_empty() && !has_no_parameters(function) =>
+                if !other_args.is_empty() || function_has_only_simple_parameters(function) =>
             {
                 grouped_arg.cache_function_body(f);
             }
@@ -604,7 +605,7 @@ impl Format<JsFormatContext> for FormatGroupedLastArgument<'_> {
         // to remove any soft line breaks.
         match element.node()? {
             AnyJsCallArgument::AnyJsExpression(JsFunctionExpression(function))
-                if !self.is_only && !has_no_parameters(function) =>
+                if !self.is_only || function_has_only_simple_parameters(function) =>
             {
                 with_token_tracking_disabled(f, |f| {
                     write!(
@@ -667,13 +668,10 @@ fn with_token_tracking_disabled<F: FnOnce(&mut JsFormatter) -> R, R>(
     result
 }
 
-/// Tests if `expression` has an empty parameters list.
-fn has_no_parameters(expression: &JsFunctionExpression) -> bool {
-    match expression.parameters() {
-        // Use default formatting for expressions without parameters, will return `Err` anyway
-        Err(_) => true,
-        Ok(parameters) => parameters.items().is_empty(),
-    }
+fn function_has_only_simple_parameters(expression: &JsFunctionExpression) -> bool {
+    expression.parameters().map_or(true, |parameters| {
+        has_only_simple_parameters(&parameters, false)
+    })
 }
 
 /// Helper for formatting a grouped call argument (see [should_group_first_argument] and [should_group_last_argument]).